Added support for ObjectID as a property type
Properties can now be defined as referring to ObjectId's.

"myProperty": { "type": "string", "format": "objectid" }

Properties defined like this will treated like actual ObjectIDs internally (and not just stored as strings).

The exception is that that if used with "mixed type" property (i.e. in conjunction with "oneOf" or "anyOf" in a schema) it will still be validated but it will be stored in the database as a string rather than an ObjectID.

For example, in this case either an object matching the "Person" schema or a string that is a valid ObjectID is permitted but in the case of an ObjectID it will be stored (incorrectly) as string internally.

"myProperty": {
  "oneOf": [
     {  "type": "string", "format": "objectid" },
     {  $ref: "Person.json }
  ]
}

Additionally, error messages when creating and updating have been improved and the README has been updated.