Changing dependancy back to json-schema-ref-parser
Thanks to @BigstickCarpet giving me the heads up on where I was going wrong with json-schema-ref-parser (and consequently why the fork is unnecessary).

I was passing an object when I should have been passing the path to the file, so it knew the relative path to load references from. When I was passing an object it looks like it was able to pull some refs from the in-memory cache but other refs to schemas it hadn't loaded yet were failing because it had no way of knowing where they had been loaded from.