Search endpoint now follows JSON API spec
The syntax for searching is now designed to follow the JSON API spec.

You can now search for any root level field names and specify one or more sort options (with sorting in both forward and reverse on each field).

Searching fields in nested objects is not supported at present.

Note: This is a breaking change as the previous search end point (/entity/search) is no longer supported, so the major version is being incremented.